Move the init command
