+++ internal/cmd/ini/cmd.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"errors"
