+++ internal/cmd/ini/common.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"fmt"
