+++ internal/cmd/ini/common_test.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"errors"
