+++ internal/cmd/ini/initializer.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"context"
