+++ internal/cmd/ini/initializer_test.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"bytes"
