+++ internal/cmd/ini/xcuitest.go
@@ -1,4 +1,4 @@
-package init
+package ini
 
 import (
 	"fmt"
